feature: Add prepare for segue method
Add prepare for segue method so I can transfer the information from one
view controller to another view controller doing a mix between coding
and using the storyboard to connect my views.

Add the functionality for when the user choose scissor as an option code
the if statement in viewWillAppear in result controller.